Adjust the code that updates the turn order UI to be more robust.
It was reported that the turn order UI text could get mangled, replacing
"Fast Power Phase" with "st Power Pha". I suspect this is because the code
to remove >> << from around the phase was being run twice, removing two more
characters from each end. Changing it to find-and-replace the >> << out
should make it robust against this.