Adjust SetupChecker panels to use the upper left corner as the point where they are positioned from.
Also, this adjusts the heights to reflect how much space is available before
running into an obstacle, and adjust the widths to match the actual width of
the contents.

While working on switching the gain a spirit functionality to use custom UI
instead of classical, I wanted to temporarily add a button to gain a spirit
near the player area. It took me a bunch of trial and error to figure out what
position to use, since the position and heights of the panels had only a vague
relation to where the visible bits of the panel were. This is because the
panels were positioned by their centers, but their height was up to about 3
times that of the contents, which resulted in the specified position being
somewhat below the visible contents.

This shouldn't affect that actual size or position of anything on screen.