Fix a potential bug with custom adversaries and the invader board UI.
The aidBoard getStage() function, when taking a substring of an
invader card's lua script in order to find that invader card's stage,
would take the entire script from the end of "cardInvaderStage="
to the end of the script. This required the definition of
cardInvaderStage to be the final line of the script in order to work
properly, and thus could break on custom adversaries with custom
invader cards that include extra scripting. This is easily fixed by
only taking one character (stage numbers are always one digit), instead
of all characters to the end of the script.