Fix a bug where spirit-specific setup items didn't work if moved to the wrong spirit first.
Namely, Days That Never Were, Deep Slumber, Sunshine, and Immense.
If the object was moved to the wrong spirit's play area first,
it would broadcast the message that it was the wrong spirit,
but the object would still have its setupComplete variable marked
as true. as such, explicitly set setupComplete to false if setup
fails due to being in the wrong spirit's area.