Fix a bug where objects' doSetup() functions could be called again after save/reload.
Fix #113.
This could occur if the objects were picked up a moved within the player area.
This was because the setupComplete variable was stored in a way that did
not persist across save/reload. Store it in script state instead to fix that.
This also require slight changes to Deep Slumber, Pour Down Power Across the Island,
and Ocean's Drowning bag, to prevent their onSave() functions overwriting the script state.