Fix a bug with the playtest expansions list when loading a save.
SetupChecker's updatePlaytestExpansionList() function expects
the Global-style table of expansions (mapping names to booleans),
not the SetupChecker-style table of expansions (mapping names to
GUIDs). SetupChecker's onLoad() gives it the SetupChecker-style
table of expansions, and thus it would treat all expansions as
being enabled. Fix this by having it not take an argument at all,
and just fetch the table of expansions from Global itself.