Cleanly broadcast an error when gaining over six cards.
It would throw an error and leave scriptWorkingCardC set to true,
so that all future card gain was impossible. There isn't any room
for more than six cards, so we can't make it work properly. But
it can at least broadcast an error message (instead of throwing
an actual Lua error) and exit cleanly.