Replace nonTokenElements with a check in aidBoard's scanElements().
To avoid double-counting element markers in player areas for the
aid board, each player area's countItems() loop counts non-token
elements separately. This makes it difficult to implement a hook
to modify elements, as both counting methods, countItems() and
aidBoard's scanElements(), have to be handled separately. To make
this easier, shift the onus for avoiding double-counting to the
aid board, removing the need for tracking non-token elements.
This likely comes with a performance penalty, as now scanElements()
must call getZones() on each element token. But scanElements() runs
only every two seconds, and getZones() isn't a raycast, so this
shouldn't be too bad.