Fix the size and position of the reminder mask.
The previous mask image was made by cropping one of the existing
spirit masks to the size of the image content. This doesn't work
properly because the origin point of the reminder object doesn't
correspond correctly to the centre of the mask image. To correct
this, regenerate the mask image from the reminder object in
Blender. Load the object in Blender, entered edit mode, ensure
the view is an orthographic projection from above centred on the
object's origin, then press "U" and "Project From View", making
sure *not* to "Project From View (Bounds)" This produces a UV map
centred on the object's origin. Export the UV map as an image,
and use an image editor to convert it to the correct
black-and-transparent image.