Preserve hand GUIDs when swapping colours.
Changing the colour of a hand requires destructing and respawning
it. Previously, the new hand was being spawned one line of code
before the old hand was destructed. As such, the GUID of the hand
was still occupied by the old copy of the hand. Swapping these
lines, to destruct the old hand before spawning the new, allows
the new hand to use the GUID of the old hand, preserving GUID.