Refactor ready helper code to improve encapsulation.
aidBoard's playerReadyGuids table didn't *just* store GUIDs, but
the Global code also maintained a set of associated colours in
there. This is a bad violation of encapsulation, and it's very
difficult to tell, just reading the aidBoard code, where all the
stored colours come from. Fix this by having Global never touch
playerReadyGuids, and having aidBoard derive the required
information directly from Global's seatTables variable. In doing
so, fix #154, as information is no longer stored in
playerReadyGuids that will be lost on save and reload.