Remove a redundant loop in getReminderImageAttributes().
We only wanted to access the first element of the table (as the table
should only have one element), but the isn't indexed by integers,
so we don't know what the key will be. I was using a loop over pairs()
and then breaking after one loop, but the linter rightly complained
that that was an ugly way of doing things. Using next() is better.