Add a nil check for script_state in getReminderLocation().
The normal value for an unset script_state is the empty string,
but single playing cards from the standard deck of 52, inexplicably,
seem to have nil as their script state. I need to check that the
code doesn't error on cards that don't have a CustomDeck in their
data, which means I need it to not error on playing cards, hence
this check.